+++ railties/lib/rails/generators/rails/controller/controller_generator.rb
@@ -11,7 +11,7 @@ module Rails
       check_class_collision suffix: "Controller"
 
       def create_controller_files
-        template "controller.rb", File.join("app/controllers", class_path, "#{file_name}_controller.rb")
+        primary_template "controller.rb", File.join("app/controllers", class_path, "#{file_name}_controller.rb")
       end
 
       def add_routes