+++ railties/lib/rails/generators/rails/task/task_generator.rb
@@ -6,7 +6,7 @@ module Rails
       argument :actions, type: :array, default: [], banner: "action action"
 
       def create_task_files
-        template "task.rb", File.join("lib/tasks", "#{file_name}.rake")
+        primary_template "task.rb", File.join("lib/tasks", "#{file_name}.rake")
       end
     end
   end